move repetitive functions to fetchProc.js
