+++ app/javascript/tapp/fetch.js
@@ -1,66 +1,16 @@
 import React from 'react';
 import { appState } from './appState.js';
+import * as fetchProc from '../fetchProc.js';
 
 /* General helpers */
 
-// display a general error text
-function msgFailure(text) {
-    appState.alert('<b>Action Failed:</b> ' + text);
-    return Promise.reject();
-}
+const msgFailure = (text) => fetchProc.msgFailure(text, appState);
+const respFailure = (resp) => fetchProc.respFailure(resp, appState);
 
-// parse a response into an error message
-function respFailure(resp) {
-    appState.alert('<b>Action Failed</b> ' + resp.url + ': ' + resp.statusText);
-    return Promise.reject();
-}
-
-function fetchHelper(URL, init) {
-    return fetch(URL, init).catch(function(error) {
-        appState.alert('<b>' + init.method + ' ' + URL + '</b> Network error: ' + error);
-        return Promise.reject(error);
-    });
-}
-
-function getHelper(URL) {
-    return fetchHelper(URL, {
-        headers: {
-            Accept: 'application/json',
-        },
-        method: 'GET',
-        credentials: 'include', // This line is crucial in any fetch because it is needed to work with Shibboleth in production
-    });
-}
-
-function postHelper(URL, body) {
-    return fetchHelper(URL, {
-        headers: {
-            Accept: 'application/json',
-            'Content-Type': 'application/json; charset=utf-8',
-        },
-        method: 'POST',
-        body: JSON.stringify(body),
-        credentials: 'include',
-    });
-}
-
-function deleteHelper(URL) {
-    return fetchHelper(URL, {
-        method: 'DELETE',
-        credentials: 'include',
-    });
-}
-
-function putHelper(URL, body) {
-    return fetchHelper(URL, {
-        headers: {
-            'Content-Type': 'application/json; charset=utf-8',
-        },
-        method: 'PUT',
-        body: JSON.stringify(body),
-        credentials: 'include',
-    });
-}
+const getHelper = (URL) => fetchProc.getHelper(URL, appState);
+const postHelper = (URL, body) => fetchProc.postHelper(URL, body, appState);
+const deleteHelper = (URL) => fetchProc.deleteHelper(URL, appState);
+const putHelper = (URL, body) => fetchProc.putHelper(URL, body, appState);
 
 /* Resource GETters */
 
@@ -76,189 +26,29 @@ const getResource = (route, onSuccess, dataName, setData) =>
 
 
 const getSessions = () => getResource(
-  '/sessions', onFetchSessionsSuccess, 'sessions', appState.setSessionsList);
+  '/sessions', fetchProc.onFetchSessionsSuccess, 'sessions', appState.setSessionsList);
 
 const getApplicants = () => getResource(
-  '/applicants', onFetchApplicantsSuccess, 'applicants', appState.setApplicantsList);
+  '/applicants', fetchProc.onFetchApplicantsSuccess, 'applicants', appState.setApplicantsList);
 
 const getApplications = () => getResource(
-  '/applications', onFetchApplicationsSuccess, 'applications',  appState.setApplicationsList);
+  '/applications', fetchProc.onFetchApplicationsSuccess, 'applications',  appState.setApplicationsList);
 
 const getCourses = () => getResource(
-  '/positions', onFetchCoursesSuccess, 'courses', appState.setCoursesList);
+  '/positions', fetchProc.onFetchTappCoursesSuccess, 'courses', appState.setCoursesList);
 
 const getAssignments = () => getResource(
-  '/assignments', onFetchAssignmentsSuccess, 'assignments', appState.setAssignmentsList);
+  '/assignments', fetchProc.onFetchAssignmentsSuccess, 'assignments', appState.setAssignmentsList);
 
 const getInstructors = () => getResource(
-  '/instructors', onFetchInstructorsSuccess, 'instructors', appState.setInstructorsList);
+  '/instructors', fetchProc.onFetchInstructorsSuccess, 'instructors', appState.setInstructorsList);
 
 
 /* Success callbacks for resource GETters */
-function onFetchSessionsSuccess(resp) {
-    let sessions = {};
-
-    resp.forEach(session => {
-        sessions[session.id] = {
-            year: session.year,
-            semester: session.semester,
-            pay: session.pay,
-        };
-    });
-
-    return sessions;
-}
-
-function onFetchApplicantsSuccess(resp) {
-    let applicants = {};
-
-    resp.forEach(app => {
-        applicants[app.id] = {
-            lastName: app.last_name,
-            firstName: app.first_name,
-            utorid: app.utorid,
-            email: app.email,
-            phone: app.phone,
-            studentNumber: app.student_number,
-            address: app.address,
-            dept: app.dept,
-            program: (function(id) {
-                switch (id) {
-                    case '1PHD':
-                        return 'PhD';
-                    case '2Msc':
-                        return 'MSc';
-                    case '3MScAC':
-                        return 'MScAC';
-                    case '4MASc':
-                        return 'MASc';
-                    case '5MEng':
-                        return 'MEng';
-                    case '6Other':
-                        return 'OG';
-                    case '7PDF':
-                        return 'PostDoc';
-                    case '8UG':
-                        return 'UG';
-                    default:
-                        return 'Other';
-                }
-            })(app.program_id),
-            year: app.yip,
-            notes: app.commentary,
-            fullTime: app.full_time == 'Y',
-        };
-    });
-
-    return applicants;
-}
-
-function onFetchApplicationsSuccess(resp) {
-    let applications = {},
-        newApp;
-
-    resp.forEach(app => {
-        newApp = {
-            taTraining: app.ta_training == 'Y',
-            academicAccess: app.access_acad_history == 'Y',
-            prefs: (function(prefs) {
-                return prefs.map(pref => ({
-                    positionId: pref.position_id,
-                    preferred: pref.rank == 1,
-                }));
-            })(app.preferences),
-            rawPrefs: app.raw_prefs,
-            exp: app.ta_experience,
-            qual: app.academic_qualifications,
-            skills: app.technical_skills,
-            avail: app.availability,
-            other: app.other_info,
-            specialNeeds: app.special_needs,
-        };
-
-        if (applications[app.applicant_id]) {
-            applications[app.applicant_id].push(newApp);
-        } else {
-            applications[app.applicant_id] = [newApp];
-        }
-    });
-
-    return applications;
-}
-
-function onFetchCoursesSuccess(resp) {
-    let courses = {};
-
-    resp.forEach(course => {
-        courses[course.id] = {
-            round: course.round_id,
-            name: course.course_name,
-            code: course.position,
-            campus: (function(code) {
-                switch (code) {
-                    case 1:
-                        return 'St. George';
-                    case 3:
-                        return 'Scarborough';
-                    case 5:
-                        return 'Mississauga';
-                    default:
-                        return 'Other';
-                }
-            })(course.campus_code),
-            instructors: course.instructors.map(instr => instr.id),
-            estimatedPositions: course.estimated_count,
-            estimatedEnrol: course.current_enrolment,
-            positionHours: course.hours,
-            cap: course.cap_enrolment,
-            waitlist: course.num_waitlisted,
-            qual: course.qualifications,
-            resp: course.duties,
-            startDate: course.start_date,
-            endDate: course.end_date,
-        };
-    });
-
-    return courses;
-}
-
-function onFetchAssignmentsSuccess(resp) {
-    let assignments = {},
-        assignmentCounts = {},
-        count,
-        newAss;
-
-    resp.forEach(ass => {
-        newAss = {
-            id: ass.id,
-            positionId: ass.position_id,
-            hours: ass.hours,
-            locked: ass.export_date != null,
-        };
-
-        if (assignments[ass.applicant_id]) {
-            assignments[ass.applicant_id].push(newAss);
-        } else {
-            assignments[ass.applicant_id] = [newAss];
-        }
-    });
-
-    return assignments;
-}
-
-function onFetchInstructorsSuccess(resp) {
-    let instructors = {};
-
-    resp.forEach(instr => {
-        instructors[instr.id] = instr.name;
-    });
-
-    return instructors;
-}
 
 /* Function to GET all resources */
 
-function fetchAll() {
+export const fetchAll = () => {
     getApplicants();
     getApplications();
     getAssignments();
@@ -269,7 +59,7 @@ function fetchAll() {
 /* Task-specific resource modifiers */
 
 // create a new assignment
-function postAssignment(applicant, course, hours) {
+export const postAssignment = (applicant, course, hours) => {
     postHelper('/applicants/' + applicant + '/assignments', {
         position_id: course,
         hours: hours,
@@ -281,7 +71,7 @@ function postAssignment(applicant, course, hours) {
 }
 
 // remove an assignment
-function deleteAssignment(applicant, assignment) {
+export const deleteAssignment = (applicant, assignment) => {
     deleteHelper('/applicants/' + applicant + '/assignments/' + assignment)
         .then(resp => (resp.ok ? resp : respFailure))
         .then(() => {
@@ -290,7 +80,7 @@ function deleteAssignment(applicant, assignment) {
 }
 
 // add/update the notes for an applicant
-function noteApplicant(applicant, notes) {
+export const noteApplicant = (applicant, notes) => {
     putHelper('/applicants/' + applicant, { commentary: notes })
         .then(resp => (resp.ok ? resp : respFailure))
         .then(() => {
@@ -299,7 +89,7 @@ function noteApplicant(applicant, notes) {
 }
 
 // update the number of hours for an assignment
-function updateAssignmentHours(applicant, assignment, hours) {
+export const updateAssignmentHours = (applicant, assignment, hours) => {
     putHelper('/applicants/' + applicant + '/assignments/' + assignment, {
         hours: hours,
     })
@@ -310,7 +100,7 @@ function updateAssignmentHours(applicant, assignment, hours) {
 }
 
 // update attribute(s) of a course
-function updateCourse(courseId, data) {
+export const updateCourse = (courseId, data) => {
     putHelper('/positions/' + courseId, data)
         .then(resp => (resp.ok ? resp : respFailure))
         .then(() => {
@@ -319,7 +109,7 @@ function updateCourse(courseId, data) {
 }
 
 // send CHASS data
-function importChass(data, year, semester) {
+export const importChass = (data, year, semester) => {
     appState.setImporting(true);
 
     postHelper('/import/chass', {
@@ -357,7 +147,7 @@ function importChass(data, year, semester) {
 }
 
 // send enrolment data
-function importInstructors(data) {
+export const importInstructors = (data) => {
     appState.setImporting(true);
 
     postHelper('/import/instructors', { instructor_data: data })
@@ -375,7 +165,7 @@ function importInstructors(data) {
 }
 
 // send enrolment data
-function importEnrolment(data) {
+export const importEnrolment = (data) => {
     appState.setImporting(true);
 
     postHelper('/import/enrolment', { enrolment_data: data })
@@ -393,7 +183,7 @@ function importEnrolment(data) {
 }
 
 // unlock a single assignment
-function unlockAssignment(applicant, assignment) {
+export const unlockAssignment = (applicant, assignment) => {
     putHelper('/applicants/' + applicant + '/assignments/' + assignment, {
         export_date: null,
     })
@@ -404,7 +194,7 @@ function unlockAssignment(applicant, assignment) {
 }
 
 // export offers from CHASS (locking the corresponding assignments)
-function exportOffers(round) {
+export const exportOffers = (round) => {
     let filename;
     let exportPromise = getHelper('/export/chass/' + round).then(
         resp => (resp.ok ? resp : respFailure)
@@ -437,7 +227,7 @@ function exportOffers(round) {
 
 // get current user role and username
 // if we are in development, set the current user name to a special value
-function fetchAuth() {
+export const fetchAuth = () => {
     getHelper('/roles')
         .then(resp => (resp.ok ? resp.json().catch(msgFailure) : respFailure))
         .then(resp => {
@@ -456,7 +246,7 @@ function fetchAuth() {
         });
 }
 
-function emailAssignments(code, round, key){
+export const emailAssignments = (code, round, key) => {
     postHelper('/email-assignments', {
         position: code,
         round_id: round,
@@ -478,19 +268,3 @@ function emailAssignments(code, round, key){
       }
     });
 }
-
-export {
-    fetchAll,
-    postAssignment,
-    deleteAssignment,
-    updateAssignmentHours,
-    updateCourse,
-    noteApplicant,
-    importChass,
-    importEnrolment,
-    importInstructors,
-    unlockAssignment,
-    exportOffers,
-    fetchAuth,
-    emailAssignments,
-};
